Initial commit
Please have in mind that this was my first big unity videogame. Some parts of the code are not properly done according to nice performance because I was just starting. I worked on that later on.

This code is not cleaned and there are many lines that do nothing. Besides, I know that making variables public is not quite correct. I've worked on that on my new projects.

Thanks for taking the time to take a look at my code